Revert "better titles"
This reverts commit 632689be72040742468632cd76ff8f607789bc84.
