+++ src/pages/SignUp.tsx
@@ -1,16 +1,30 @@
 import { useEffect, useState, useCallback } from "react";
-import { getApiV1UserTerms, postApiV1UserTerms } from "@/client";
+import { postApiV1AuthSignup, postApiV1AuthSignupOtp, getApiV1UserTerms, postApiV1UserTerms } from "@/client";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Button } from "@/components/ui/button";
 import { StandardAlert } from "@/components/ui/standard-alert";
+import { OtpInput } from "@/components/otpInput";
 import { useAuth } from "@/context/AuthContext";
 import { useUser } from "@/context/UserContext";
 import { useNavigate } from "react-router-dom";
 import { extractErrorMessage } from "@/utils/errorHelpers";
-import EmailVerificationStep from "@/components/safe-deployment/EmailVerificationStep";
+import { userTerms, type UserTermsTypeFromApi } from "@/constants";
+
+enum ScreenStep {
+  EmailAndTos = "email-and-tos",
+  OtpVerification = "otp-verification",
+}
 
 export const SignUpRoute = () => {
   const { updateJwt, updateClient } = useAuth();
   const { isUserSignedUp } = useUser();
+  const [step, setStep] = useState<ScreenStep>(ScreenStep.EmailAndTos);
+  const [email, setEmail] = useState("");
+  const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState("");
+  const [otp, setOtp] = useState("");
+  const [isAcceptedTos, setIsAcceptedTos] = useState(false);
   const navigate = useNavigate();
 
   useEffect(() => {
@@ -64,42 +78,136 @@ export const SignUpRoute = () => {
     }
   }, []);
 
-  const handleEmailVerificationComplete = useCallback(
-    async (token?: string) => {
-      if (!token) return;
+  const handleSubmitOtpRequest = useCallback(
+    async (e: React.FormEvent) => {
+      e.preventDefault();
+      setIsLoading(true);
+      setError("");
+      setOtp("");
 
       try {
+        const { error, data } = await postApiV1AuthSignupOtp({
+          body: { email },
+        });
+        if (error) {
+          const message = extractErrorMessage(error, "unknown");
+          setError(`Error returned while requesting the OTP: ${message}`);
+          console.error("Error returned while requesting the OTP", error);
+        }
+
+        data?.ok && setStep(ScreenStep.OtpVerification);
+      } catch (err) {
+        const message = extractErrorMessage(err, "unknown");
+        setError(`Error while requesting the OTP: ${message}`);
+        console.error("Error requesting OTP", err);
+      } finally {
+        setIsLoading(false);
+      }
+    },
+    [email],
+  );
+
+  const handleOtpSubmit = useCallback(
+    async (e: React.FormEvent) => {
+      e.preventDefault();
+      setIsLoading(true);
+      setError("");
+
+      try {
+        const { error, data } = await postApiV1AuthSignup({
+          body: { authEmail: email, otp },
+        });
+
+        if (error || !data) {
+          const message = extractErrorMessage(error, "unknown");
+          setError(`Error returned while signing up: ${message}`);
+          console.error("Error returned while signing up", error);
+          return;
+        }
+
         // unless we update the client with the new JWT, it will not be used for subsequent requests
         // and we will not be able to accept user terms
-        updateJwt(token);
-        updateClient(token);
+        updateJwt(data.token);
+        updateClient(data.token);
 
         await acceptAllUserTerms();
         navigate("/kyc");
       } catch (err) {
         const message = extractErrorMessage(err, "unknown");
-        setError(`Error while completing signup: ${message}`);
-        console.error("Error while completing signup", err);
+        setError(`Error while signing up: ${message}`);
+        console.error("Error while signing up", err);
+      } finally {
+        setIsLoading(false);
       }
     },
-    [updateJwt, updateClient, acceptAllUserTerms, navigate],
+    [email, otp, updateJwt, acceptAllUserTerms, updateClient, navigate],
   );
 
   return (
     <div className="grid grid-cols-6 gap-4 h-full mt-4">
-      {error && (
-        <div className="col-span-6 lg:col-start-2 lg:col-span-4 mx-4 lg:mx-0">
-          <StandardAlert variant="destructive" title="Error" description={error} />
-        </div>
-      )}
-      <EmailVerificationStep
-        onComplete={handleEmailVerificationComplete}
-        setError={setError}
-        requireToS={true}
-        submitButtonText="Get code"
-        title="Sign up to Gnosis Pay"
-        description="Type your email to receive a one time code."
-      />
+      <div className="col-span-6 lg:col-start-2 lg:col-span-4 mx-4 lg:mx-0">
+        {step === ScreenStep.EmailAndTos && (
+          <form className="space-y-4 mt-8" onSubmit={handleSubmitOtpRequest}>
+            <Label htmlFor="register-email">Type your email to receive a one time code.</Label>
+            <div className="mt-4">
+              <Input
+                className="lg:w-1/2"
+                id="register-email"
+                type="email"
+                placeholder="you@example.com"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                required
+                autoComplete="email"
+                disabled={isLoading}
+              />
+            </div>
+            <div className="flex items-center space-x-2 mt-2">
+              <input
+                id="accept-tos"
+                type="checkbox"
+                checked={isAcceptedTos}
+                onChange={(e) => setIsAcceptedTos(e.target.checked)}
+                disabled={isLoading}
+                required
+              />
+              <Label htmlFor="accept-tos" className="text-sm">
+                I have read and agree to the{" "}
+                {Object.entries(userTerms).map(([type, { url }], idx, arr) => (
+                  <span key={type}>
+                    <a href={url} target="_blank" rel="noopener noreferrer" className="underline">
+                      {userTerms[type as UserTermsTypeFromApi].title}
+                    </a>
+                    {idx < arr.length - 1 ? ", " : ""}
+                  </span>
+                ))}
+              </Label>
+            </div>
+
+            <Button type="submit" loading={isLoading} disabled={isLoading || !email || !isAcceptedTos}>
+              Get code
+            </Button>
+            {error && <StandardAlert variant="destructive" title="Error" description={error} />}
+          </form>
+        )}
+        {step === ScreenStep.OtpVerification && (
+          <form className="space-y-4 mt-8" onSubmit={handleOtpSubmit}>
+            <Label htmlFor="otp">Enter the 6-digit code sent to your email</Label>
+            <OtpInput value={otp} onChange={setOtp} isLoading={isLoading} disabled={isLoading} />
+            {error && (
+              <Button variant={"link"} onClick={handleSubmitOtpRequest}>
+                Send new code
+              </Button>
+            )}
+            {!error && (
+              <Button type="submit" loading={isLoading} disabled={isLoading || otp.length !== 6}>
+                Verify and Sign Up
+              </Button>
+            )}
+            {error && <StandardAlert variant="destructive" title="Error" description={error} />}
+          </form>
+        )}
+      </div>
     </div>
   );
 };
